[skip ci] wip: rfox bridge refactor PoC
